Add install script
Developed to not only include setting-up the expected configuration
file in its location, but also add the program to an expected location
on the PATH for many terminal users.

Not a huge use case, but thought I would start to provide the
convenience, and we can always improve on it later.